test(vis-reg): make all-vis-reg a usable two-URL site comparison
Repair and harden the all-vis-reg A/B tool so it reliably compares any
two live YaleSites URLs (baseline vs candidate):

- Sync the page list to the 30 component test pages currently exposed
  (add directory, spotlight-landscape, spotlight-portrait, tiles; drop
  the defunct content-spotlight).
- Add load stabilization before each capture (network idle, image and
  font load, animation settle, WebKit overflow-x fix) so cold Pantheon
  sandboxes do not yield partial screenshots. The logic is extracted to
  a shared support/stabilizePage.ts helper and reused by the
  per-component vis-reg helper, removing the duplicated copy.
- Tighten maxDiffPixelRatio to 0.001 and raise the screenshot
  stabilization timeout to 20s so real deltas surface for triage while
  image-heavy pages still settle.
- Tune playwright.config.ts for slow Pantheon sandboxes (90s navigation,
  180s per-test budget, 2 retries, single worker).
- Rework compareSites.sh: default to the dev/v2230 sandboxes, warm them
  first, emit a JSON report, tolerate the expected non-zero exits so the
  HTML report still opens, and add --reuse-baseline to skip recapturing
  an existing baseline.
- Gitignore the regenerated all-vis-reg baselines; they are working data,
  not a committed reference (unlike the per-component golden baselines).
