+++ tests/all-vis-reg.spec.ts
@@ -1,40 +1,53 @@
-// This is an attempt to come up with a visual regression test utilizing Playwright Test
-//  that can be used to compare a source (expected) website that is different from the
-//  test (actual) website. The built-in `expect(page).toHaveScreenshot()` functionality
-//  only works with the same base URL.
-
-//  Unfortunately while the test below will fail a visual regression, it won't take
-//  advantage of the nice GUI inspection of the failure in the HTML report that
-//  Playwrigt Test otherwise provides if you use `toHaveScreenshot`
+// Visual regression A/B comparison between two live YaleSites environments.
+// The built-in `expect(page).toHaveScreenshot()` only diffs against a stored
+// baseline, so the workflow is: capture the baseline from the currently-released
+// site with `--update-snapshots`, then run again against the candidate site to
+// diff. See scripts/compareSites.sh for the wrapper.
+//
+// Example:
+//   rm -rf test-results snapshots/all-vis-reg.spec.ts-snapshots
+//   YALESITES_URL="https://dev-ys-yalesites-visreg-yale-edu.pantheonsite.io"  npx playwright test all-vis-reg.spec --update-snapshots
+//   YALESITES_URL="https://v2230-ys-yalesites-visreg-yale-edu.pantheonsite.io" npx playwright test all-vis-reg.spec --reporter=html,json
+//   npx playwright show-report
 import { test, expect } from "@playwright/test";
-
-/*
-rm -rf test-results; rm -rf snapshots/all-vis-reg.spec.ts-snapshots; YALESITES_URL="https://dev-yalesites-mv482.pantheonsite.io" npx playwright test all-vis-reg.spec; YALESITES_URL="https://d10-yalesites-mv482.pantheonsite.io" npx playwright test all-vis-reg.spec; npx playwright show-report
-*/
+import stabilizePage from "@support/stabilizePage";
 
 const testScreenshotForPage = (pagePath: string) => {
-  // Sleep for 5 seconds to not bombard the server
-  // setTimeout(() => {}, 5000);
-  const subPath = "component-pages-for-e2e-testing"
-  const url = `${process.env.YALESITES_URL || "http://yalesites-platform.lndo.site"}/${subPath}/${pagePath}`;
+  const subPath = "component-pages-for-e2e-testing";
+  const base = process.env.YALESITES_URL || "http://yalesites-platform.lndo.site";
+  const url = `${base}/${subPath}/${pagePath}`;
+
   return test(`should compare ${pagePath}`, async ({ page }) => {
-    // await page.waitForTimeout(5000);
-    await page.goto(url);
+    await page.goto(url, { waitUntil: "domcontentloaded" });
+
+    // Settle the page (network, images, fonts, animations) before capturing —
+    // Pantheon sandboxes are slow to warm up, so screenshots can be partial.
+    await stabilizePage(page);
+
     await expect(page).toHaveScreenshot({
       fullPage: true,
-      maxDiffPixelRatio: 0.17,
+      // Strict: surface any meaningful visual delta for manual triage rather than
+      // silently passing changes under a loose release-gate threshold.
+      maxDiffPixelRatio: 0.001,
       animations: "disabled",
+      // Give slow-loading Pantheon images time to settle so the stabilization
+      // step can capture two consecutive identical frames (default 5s is too
+      // short for image-heavy pages like the view/event listing on WebKit).
+      timeout: 20000,
     });
   });
 };
 
+// The 30 component test pages exposed under /component-pages-for-e2e-testing/
+// on both the dev and v2230 environments. Keep this list in sync with the
+// sitemap if components are added or renamed.
 testScreenshotForPage("accordion");
 testScreenshotForPage("action-banner");
 testScreenshotForPage("button-link");
 testScreenshotForPage("calendar-list");
 testScreenshotForPage("callout");
-testScreenshotForPage("content-spotlight");
 testScreenshotForPage("custom-cards");
+testScreenshotForPage("directory");
 testScreenshotForPage("divider");
 testScreenshotForPage("embed");
 testScreenshotForPage("facts-and-figures");
@@ -50,8 +63,11 @@ testScreenshotForPage("pre-built-form");
 testScreenshotForPage("quick-links");
 testScreenshotForPage("quote");
 testScreenshotForPage("quote-callout");
+testScreenshotForPage("spotlight-landscape");
+testScreenshotForPage("spotlight-portrait");
 testScreenshotForPage("tabs");
 testScreenshotForPage("text");
+testScreenshotForPage("tiles");
 testScreenshotForPage("video");
 testScreenshotForPage("view");
 testScreenshotForPage("wrapped-image");
